cdk/awscdk/bento/requirements.txt
-e .
aws-cdk.core
aws-cdk.aws_iam
configparser